+++ language/pt-BR/administrator.php
@@ -1,0 +1,55 @@
+<?php
+
+return array(
+
+	/*
+	|--------------------------------------------------------------------------
+	| Administrator Language Lines
+	|--------------------------------------------------------------------------
+	|
+	| pt-BR - Brazillian Portuguese
+	|
+	*/
+
+	'dashboard' => 'Painel',
+	'edit' => 'Editar',
+	'filters' => 'Filtros',
+	'loading' => 'Carregando...',
+	'createnew' => 'Criar novo',
+	'new' => 'Novo',
+	'viewitem' => 'Visualizar :single',
+	'id' => 'ID',
+	'uploadimage' => 'Enviar imagem',
+	'imageuploading' => 'Enviando imagem',
+	'uploadfile' => 'Enviar arquivo',
+	'fileuploading' => 'Enviando arquivo',
+	'none' => 'Nenhum',
+	'all' => 'Todos',
+	'itemsperpage' => 'itens por página',
+	'noresults' => 'Sem Resultados',
+	'backtosite' => 'Voltar para o Site',
+
+	'previous' => 'anterior',
+	'next' => 'próximo',
+
+	'close' => 'Fechar',
+	'delete' => 'Deletar',
+	'save' => 'Salvar',
+	'create' => 'Criar',
+	'cancel' => 'Cancelar',
+
+	'active' => 'Um momento...',
+	'success' => 'Sucesso!',
+	'error' => 'Ocorreu um erro ao realizar esta ação.',
+
+	'valid_columns' => "Você precisa fornecer um array válido em 'columns' em cada configuração dos models.",
+	'valid_title' => "Você precisa fornecer um título e nome singular em cada configuração dos models",
+	'valid_model' => "Você precisa fornecer uma opção válida em 'model' para cada configuração de models",
+	'valid_edit' => "Você precisa fornecer uma opção válida em 'edit_fields' para cada configuração de models",
+	'valid_menu' => "Você precisa fornecer uma opção válida em 'menu' no arquivo de configuração administrator.php",
+	'valid_config_path' => "Opção inválida em 'model_config_path' no arquivo de configuração administrator.php. O diretório precisa existir e ter permissão de leitura.",
+	'not_eloquent' => " não é um model Eloquent",
+	'storage_path_permissions' => "O caminho de storage precisa ter permissão de escrita.",
+	'valid_home_page' => "Você precisa fornecer um item de menu válido na opção 'home_page'.",
+);
+