+++ views/templates/edit.php
@@ -208,7 +208,7 @@
 	<!-- ko if: $root[$root.primaryKey]() && actions.length -->
 		<div class="custom_buttons">
 			<!-- ko foreach: actions -->
-				<!-- ko if: hasPermission -->
+				<!-- ko if: hasPermission && $root.actionPermissions[name] !== false -->
 					<input type="button" data-bind="click: function(){$root.customAction(name, messages)}, value: title,
 																	attr: {disabled: $root.freezeForm() || $root.freezeActions()}" />
 				<!-- /ko -->