+++ changelog.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 3.3.0
+- You can now define your custom action's permissions in the model config's action_permissions option
+- Returning Response or Redirect objects is now possible for admin and model configs
+- New language: Brazilian Portuguese (pt-BR)
+- Bugfix: Self-relationships weren't updating properly since the move to select2
+- Bugfix: Columns for HasMany, HasOne, and HMABT relationships now work
+
+
 ### 3.2.0
 - Added support for a file field
 - You can now choose to provide a custom dashboard or a default home page from your menu